Extract request_handler from Reddit
