+++ apraw/reddit.py
@@ -1,15 +1,12 @@
 import asyncio
 import configparser
 import os
-from datetime import datetime, timedelta
-from functools import wraps
-from typing import Any, Awaitable, Callable, Dict, List, Union
+from typing import Dict, List, Union
 
-from multidict import CIMultiDictProxy
-
-from .endpoints import API_PATH, BASE_URL
+from .endpoints import API_PATH
 from .models import (Comment, Listing, Redditor, Submission,
                      Subreddit, User, ListingGenerator, streamable)
+from .request_handler import RequestHandler
 from .utils import prepend_kind
 
 if os.path.exists('praw.ini'):
@@ -310,135 +307,3 @@ class Reddit:
             data["from_sr"] = str(from_sr)
         resp = await self.post_request(API_PATH["compose"], data=data)
         return not resp["json"]["errors"]
-
-
-class RequestHandler:
-
-    def __init__(self, user: User):
-        self.user = user
-        self.queue = []
-
-    async def get_request_headers(self) -> Dict:
-        if self.user.token_expires <= datetime.now():
-            url = "https://www.reddit.com/api/v1/access_token"
-            session = await self.user.auth_session()
-
-            headers = {
-                "Content-Type": "application/x-www-form-urlencoded",
-                "User-Agent": self.user.user_agent
-            }
-
-            resp = await session.post(url, data=self.user.password_grant, headers=headers)
-
-            async with resp:
-                if resp.status == 200:
-                    self.user.access_data = await resp.json()
-                    self.user.token_expires = datetime.now(
-                    ) + timedelta(seconds=self.user.access_data["expires_in"])
-                else:
-                    raise Exception("Invalid user data.")
-
-        return {
-            "Authorization": "{} {}".format(self.user.access_data["token_type"], self.user.access_data["access_token"]),
-            "User-Agent": self.user.user_agent
-        }
-
-    def update(self, data: CIMultiDictProxy[str]):
-        if "x-ratelimit-remaining" in data:
-            self.user.ratelimit_remaining = int(float(data["x-ratelimit-remaining"]))
-        if "x-ratelimit-used" in data:
-            self.user.ratelimit_used = int(data["x-ratelimit-used"])
-        if "x-ratelimit-reset" in data:
-            self.user.ratelimit_reset = datetime.now() + timedelta(seconds=int(data["x-ratelimit-reset"]))
-
-    async def close(self):
-        await self.user.close()
-
-    class Decorators:
-
-        @classmethod
-        def check_ratelimit(
-                cls, func: Callable[[Any], Awaitable[Any]]) -> Callable[[Any], Awaitable[Any]]:
-            @wraps(func)
-            async def execute_request(self, *args, **kwargs) -> Any:
-                id = datetime.now().strftime('%Y%m%d%H%M%S')
-                self.queue.append(id)
-
-                if self.user.ratelimit_remaining < 1:
-                    execution_time = self.user.ratelimit_reset + \
-                                     timedelta(seconds=len(self.queue))
-                    wait_time = (
-                            execution_time -
-                            datetime.now()).total_seconds()
-                    await asyncio.sleep(wait_time)
-
-                result = await func(self, *args, **kwargs)
-                self.queue.remove(id)
-                return result
-
-            return execute_request
-
-    @Decorators.check_ratelimit
-    async def get(self, endpoint: str = "", **kwargs) -> Dict:
-        kwargs = {"raw_json": 1, "api_type": "json", **kwargs}
-        params = ["{}={}".format(k, kwargs[k]) for k in kwargs]
-
-        url = BASE_URL.format(endpoint, "&".join(params))
-
-        headers = await self.get_request_headers()
-        session = await self.user.client_session()
-        resp = await session.get(url, headers=headers)
-
-        async with resp:
-            self.update(resp.headers)
-            return await resp.json()
-
-    @Decorators.check_ratelimit
-    async def delete(self, endpoint: str = "", **kwargs) -> Dict:
-        kwargs = {"raw_json": 1, "api_type": "json", **kwargs}
-        params = ["{}={}".format(k, kwargs[k]) for k in kwargs]
-
-        url = BASE_URL.format(endpoint, "&".join(params))
-
-        headers = await self.get_request_headers()
-        session = await self.user.client_session()
-        resp = await session.delete(url, headers=headers)
-
-        async with resp:
-            self.update(resp.headers)
-            return await resp.json()
-
-    @Decorators.check_ratelimit
-    async def put(self, endpoint: str = "", **kwargs) -> Dict:
-        kwargs = {"raw_json": 1, "api_type": "json", **kwargs}
-        params = ["{}={}".format(k, kwargs[k]) for k in kwargs]
-
-        url = BASE_URL.format(endpoint, "&".join(params))
-
-        headers = await self.get_request_headers()
-        session = await self.user.client_session()
-        resp = await session.delete(url, headers=headers)
-
-        async with resp:
-            self.update(resp.headers)
-            return await resp.json()
-
-    @Decorators.check_ratelimit
-    async def post(self, endpoint: str = "", url: str = "", data: Dict = None, **kwargs) -> Dict:
-        if not data:
-            data = {}
-        kwargs = {"raw_json": 1, "api_type": "json", **kwargs}
-        params = ["{}={}".format(k, kwargs[k]) for k in kwargs]
-
-        if endpoint:
-            url = BASE_URL.format(endpoint, "&".join(params))
-        elif url:
-            url = "{}?{}".format(url, "&".join(params))
-
-        headers = await self.get_request_headers()
-        session = await self.user.client_session()
-        resp = await session.post(url, data=data, headers=headers)
-
-        async with resp:
-            self.update(resp.headers)
-            return await resp.json()
